don't cleanup the path if the LB_NO_STRIP_PATH variable is present
git-svn-id: http://svn.cern.ch/guest/lhcb/LbScripts/trunk@13804 4525493e-7705-40b1-a816-d608a930855b

+++ LbScripts/LbScriptsPolicy/scripts/StripPath.csh
@@ -1,7 +1,6 @@
-
-set tmpfile = `runpy LbScriptsPolicy.PathStripper --shell=csh --mktemp -e PATH -e LD_LIBRARY_PATH -e PYTHONPATH -e HPATH `
-source $tmpfile
-rm -f $tmpfile
-unset tmpfile
-unset scriptdir
-
+if (! ${?LB_NO_STRIP_PATH} ) then
+  set tmpfile = `runpy LbScriptsPolicy.PathStripper --shell=csh --mktemp -e PATH -e LD_LIBRARY_PATH -e PYTHONPATH -e HPATH `
+  source $tmpfile
+  rm -f $tmpfile
+  unset tmpfile
+endif
