+++ LbScripts/LbScriptsPolicy/scripts/StripPath.sh
@@ -1,7 +1,6 @@
-scriptdir=$LBSCRIPTSPOLICYROOT/scripts
-
-tmpfile=`runpy LbScriptsPolicy.PathStripper --shell=sh --mktemp -e PATH -e LD_LIBRARY_PATH -e PYTHONPATH -e JOBOPTSEARCHPATH -e HPATH `
-. $tmpfile
-rm -f $tmpfile
-unset tmpfile
-unset scriptdir
+if [[ "x$LB_NO_STRIP_PATH" ==  "x" ]] ; then
+  tmpfile=`runpy LbScriptsPolicy.PathStripper --shell=sh --mktemp -e PATH -e LD_LIBRARY_PATH -e PYTHONPATH -e JOBOPTSEARCHPATH -e HPATH `
+  . $tmpfile
+  rm -f $tmpfile
+  unset tmpfile
+fi
